Add a new package risc0-binfmt
We need to drop risc0-zkvm as a dependency for risc0-build to avoid future
circular dependencies. In order to do so, we introduce a new crate,
risc0-binfmt that contains all code needed by risc0-build that used to be a
part of risc0-zkvm. This crate contains Program, MemoryImage, and SystemState
struct definitions along with several functions that operate on it.